workflows: add unified workflow mixin
* Adds a new mixin for workflow definitions which use SIP as
  data storage format. Contains a set of common functions used with
  Holding Pen.
